photo gallery

+++ src/components/Dropzone.jsx
@@ -5,13 +5,13 @@ export default function Dropzone(props) {
   const [myFiles, setMyFiles] = useState([]);
 
   const onDrop = (acceptedFiles) => {
-    props.onDrop(acceptedFiles[0]);
+    if (typeof props.onDrop !== 'undefined') {
+      props.onDrop(acceptedFiles[0]);
+    }
     setMyFiles([...acceptedFiles]);
   };
 
-  const { getRootProps, getInputProps, isDragActive } = useDropzone({
-    onDrop
-  });
+  const { getRootProps, getInputProps, isDragActive, isFileDialogActive } = useDropzone({onDrop});
 
   const files = myFiles.map((file) => {
     return (
@@ -22,18 +22,34 @@ export default function Dropzone(props) {
   });
 
   const removeAll = () => {
-    props.onRemove(null)
+    if (typeof props.onRemove !== 'undefined') {
+      props.onRemove(null)
+    }
     setMyFiles([]);
   };
 
+
+  if (props.hasOwnProperty('galleryImages')
+      && parseInt(props.galleryImages) >= parseInt(props.maxNrOfFiles)) {
+   return (
+       <section className='container sm:w-1/3 border-dashed border-2 border-gray-600 px-2 py-4'>
+         <div className='px-4 py-4 cursor-pointer bg-red-100'>
+           <p>You reached the number of files allowed!</p>
+         </div>
+       </section>
+   );
+  }
+
   return (
-    <section className='container sm:w-2/3 border-dashed border-2 border-gray-600 px-2 py-4'>
+    <section className='container sm:w-1/3 border-dashed border-2 border-gray-600 px-2 py-4'>
       <div {...getRootProps({ className: 'dropzone' })}
-           className={`px-4 py-4 ${ isDragActive ? 'bg-green-100' : ''}`}>
+           className={`px-4 py-4 cursor-pointer ${ isDragActive || isFileDialogActive ? 'bg-green-100' : ''}
+                ${typeof props.onDrop !== 'undefined' ? 'h-full' : ''}
+           `}>
         <input {...getInputProps()} />
-        <p>Upload new photo</p>
+        <p>Click to select/ Upload new photo</p>
       </div>
-      {myFiles.length > 0 && (
+      {myFiles.length > 0 && typeof props.onRemove !== 'undefined' && (
         <aside className='flex ml-4'>
           <h4 className={' mr-4'}>New Image:</h4>
           <ul>{files}</ul>
@@ -42,7 +58,7 @@ export default function Dropzone(props) {
             className='bg-red-500 text-white active:bg-red-600 ml-4 font-bold uppercase text-xs px-4 py-2 rounded shadow hover:shadow-md outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150'
             type='button'
           >
-            remove
+            Remove
           </button>
         </aside>
       )}
